agent, detection, spiking und tenacity param tabelle als nested form ohne java script

+++ db/migrate/20140710214358_create_tenacities.rb
@@ -2,7 +2,7 @@ class CreateTenacities < ActiveRecord::Migration
   def change
     create_table :tenacities do |t|
       t.integer :total_sample_number
-      t.integer :inoculum_concentration_initial
+      t.decimal :inoculum_concentration_initial
       t.string :i_concentration_unit
       t.integer :reference_id
       t.integer :user_id
